requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
discord.py==2.4.0
kubernetes==31.0.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
mcp==1.3.0
